[SPARK-48056][CONNECT][FOLLOW-UP] Scala Client re-execute plan if a SESSION_NOT_FOUND error is raised and no partial response was received
### What changes were proposed in this pull request?

This change lets a Scala Spark Connect client reattempt execution of a plan when it receives a SESSION_NOT_FOUND error from the Spark Connect service if it has not received any partial responses.

This is a Scala version of the previous fix of the same issue - https://github.com/apache/spark/pull/46297.

### Why are the changes needed?

Spark Connect clients often get a spurious error from the Spark Connect service if the service is busy or the network is congested. This error leads to a situation where the client immediately attempts to reattach without the service being aware of the client; this leads to a query failure.

### Does this PR introduce _any_ user-facing change?

Prevoiusly, a Scala Spark Connect client would fail with the error code "INVALID_HANDLE.SESSION_NOT_FOUND" in the very first attempt to make a request to the service, but with this change, the client will automatically retry.

### How was this patch tested?

Attached unit test.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #46971 from changgyoopark-db/SPARK-48056.

Authored-by: Changgyoo Park <changgyoo.park@databricks.com>
Signed-off-by: Ruifeng Zheng <ruifengz@apache.org>