[SPARK-48587][VARIANT] Avoid storage amplification when accessing a sub-Variant
### What changes were proposed in this pull request?

Previously, if `variant_get` returned a Variant or a nested type containing Variant, we would just return the slice of the value along with the full metadata. Typically, most of the metadata is relevant to other parts of the original Variant value, and is not needed, so the resulting binary is larger than needed. This can be very expensive if the value is written to disk (e.g. parquet or shuffle file).

### Why are the changes needed?

Avoid unnecessarily large Variant values in memory and on disk.

### Does this PR introduce _any_ user-facing change?

No, the resulting Variant is logically the same as before, only storage size should change.

### How was this patch tested?

Unit tests.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #46941 from cashmand/SPARK-48587.

Authored-by: cashmand <david.cashman@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>