[SQL][TEST][FOLLOWUP] Re-run collation benchmark (NonASCII)
### What changes were proposed in this pull request?
Following up on https://github.com/apache/spark/pull/47030, re-running the collation benchmark for NonASCII.

### Why are the changes needed?
We've changed the meaning of LCASE collation in Spark, and also modified how equality checks / hashing/ expressions work with this collation, so we need to re-run the benchmarks and identify areas of improvement.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
Existing tests.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #47054 from uros-db/collation-benchmarks-nonascii.

Authored-by: Uros Bojanic <157381213+uros-db@users.noreply.github.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>