[SPARK-48718][SQL] Handle and fix the case when deserializer in cogroup is resolved during application of DeduplicateRelation rule
### What changes were proposed in this pull request?
A followup for https://github.com/apache/spark/pull/41554/files.
Handle the case when the deserializer in cogroup is resolved when applying DeduplicateRelation rule. Otherwise, it will throw an uncastable error.
See the added test case as an example.

### Why are the changes needed?
Fix a bug introduced in a previous commit.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
Add a new test case.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #47091 from anchovYu/fix-cogroup-dedup-rel.

Lead-authored-by: Xinyi Yu <xinyi.yu@databricks.com>
Co-authored-by: Wenchen Fan <cloud0fan@gmail.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>