[SPARK-48843] Prevent infinite loop with BindParameters
### What changes were proposed in this pull request?

In order to resolve the named parameters on the subtree, BindParameters recurses into the subtrees and tries to match the pattern with the named parameters. If there's no named parameter in the current level, the rule tries to return the unchanged plan. However, instead of returning the current plan object, the rule always returns the captured root plan node, leading into the infinite recursion.

### Why are the changes needed?

Infinite recursion with the named parameters and the global limit.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Added unit tests.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #47271 from nemanja-boric-databricks/fix-bind.

Lead-authored-by: Nemanja Boric <nemanja.boric@databricks.com>
Co-authored-by: Wenchen Fan <cloud0fan@gmail.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>