[SPARK-48844][SQL] USE INVALID_EMPTY_LOCATION instead of UNSUPPORTED_DATASOURCE_FOR_DIRECT_QUERY when path is empty
### What changes were proposed in this pull request?

When running sql on valid datasource files directly, if the given path is an empty string, we currently report UNSUPPORTED_DATASOURCE_FOR_DIRECT_QUERY, which claims the datasource is invalid. The reason is that the `hadoop.Path` class can not be constructed with empty strings and we wrap `IAE` with UNSUPPORTED_DATASOURCE_FOR_DIRECT_QUERY.

In this PR, we check the path ahead to avoid this ambiguous error message

### Why are the changes needed?

trivial bugfix, although this error rarely occurs in REPL environments but might still get a chance to happen when using the query with string interpolation.

### Does this PR introduce _any_ user-facing change?

Yes, different error class

### How was this patch tested?

new tests
### Was this patch authored or co-authored using generative AI tooling?
no

Closes #47267 from yaooqinn/SPARK-48844.

Authored-by: Kent Yao <yao@apache.org>
Signed-off-by: Dongjoon Hyun <dhyun@apple.com>