[SPARK-48775][SQL][STS] Replace SQLContext with SparkSession in STS
### What changes were proposed in this pull request?

Remove the exposed `SQLContext` which was added in SPARK-46575. And migrate STS internal used `SQLContext` to `SparkSession`.

### Why are the changes needed?

`SQLContext` is not recommended since Spark 2.0, the suggested replacement is `SparkSession`. We should avoid exposing the deprecated class to Developer API in new versions.

### Does this PR introduce _any_ user-facing change?

No. It touched the Developer API added in SPARK-46575, but is not released yet.

### How was this patch tested?

Pass GHA, and `dev/mima` (not breaking changes involved)

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #47176 from pan3793/SPARK-48775.

Lead-authored-by: Cheng Pan <chengpan@apache.org>
Co-authored-by: Cheng Pan <pan3793@gmail.com>
Signed-off-by: Dongjoon Hyun <dhyun@apple.com>