[SPARK-36680][SQL][FOLLOWUP] Files with options should be put into resolveDataSource function
### What changes were proposed in this pull request?

When reading csv, json and other files, pass the options parameter to the rules.resolveDataSource method to make the options parameter effective.

This is a bug fix for [#46707](https://github.com/apache/spark/pull/46707) szehon-ho

### Why are the changes needed?

For the following SQL, the options parameter passed in does not take effect. This is because the rules.resolveDataSource method does not pass the options parameter during the datasource construction process

```
 SELECT * FROM csv.`/test/data.csv` WITH (`header` = true, 'delimiter' = '|')
 ```

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Unit test in SQLQuerySuite

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #47370 from logze/hint-options.

Authored-by: lizongze <lizongze@xiaomi.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>