[SPARK-48929] Fix view internal error and clean up parser exception context
### What changes were proposed in this pull request?

Replace INVALID_VIEW_TEXT (SQLSTATE: XX000) error-condition which is forced when view expansion fails to parse with the underlying error.
The reason being that the underlying error has important information AND it is much more likely that view expansion fails because of a behavior change in the parser than some nefarious action or simple corruption.
For example the recent removal of `!` and `NOT` equivalence fails old view on on Spark 4.0 if they exploit the unsupported syntax.

As part of this change we also include the Query context summary in the ParserException processing.
This will expose the view name along side the view query error message.

### Why are the changes needed?

Better error handling improving ability of customers to do root cause analysis.

### Does this PR introduce _any_ user-facing change?

We are now giving more error context in case of parer errors and specifically in case of parser errors in view expansion

### How was this patch tested?

Existing QA

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #47405 from srielau/SPARK-48929-Fix-view-intrenal-error.

Authored-by: Serge Rielau <serge@rielau.com>
Signed-off-by: Gengliang Wang <gengliang@apache.org>