[SPARK-48986][CONNECT][SQL] Add ColumnNode Intermediate Representation
### What changes were proposed in this pull request?
This PR introduces an intermediate representation for Column operations. It also adds a converter from this IR to Catalyst Expression. This is a first step in sharing Column API between Classic and Connect. It is not integrated with any of the pre-existing code base.

### Why are the changes needed?
We want to share the Scala Column API between the Classic and Connect Scala DataFrame API implementations. For this we need to decouple the Column API from Catalyst.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
I added a test suite that tests the conversion from `ColumnNode` to `Expression`.

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #47466 from hvanhovell/SPARK-48986.

Authored-by: Herman van Hovell <herman@databricks.com>
Signed-off-by: Herman van Hovell <herman@databricks.com>