[SPARK-49125][SQL] Allow duplicated column names in CSV writing
### What changes were proposed in this pull request?

In file source writing, we disallow duplicated column names in the input query for all formats, because most formats don't do well with duplicate columns. However, this is not a good decision as long as there are formats allowing duplicated column names, and CSV is one of them.

This PR improves the `FileFormat` API to indicate if the format allows duplicated column names or not, and only perform the duplicated name check for formats that don't allow it.

### Why are the changes needed?

enable more use cases for the CSV data source

### Does this PR introduce _any_ user-facing change?

Yes, now users can write to CSV with duplicated column names.

### How was this patch tested?

new tests

### Was this patch authored or co-authored using generative AI tooling?

no

Closes #47633 from cloud-fan/csv.

Authored-by: Wenchen Fan <wenchen@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>