[SPARK-49142][CONNECT][PYTHON] Follow up to revert proto to string performance cost
### What changes were proposed in this pull request?

In a previous commit 923d06f9, log level in the connect client
was lowered from info to debug. As part of that, the explicit
check for logging level was removed.

Revert this change and put the logging level check back.

### Why are the changes needed?

The explicit check was there so as to not convert the proto
to a string, done during string interpolation, when not at the
expected log level. Conversion from proto to string can be
expensive for some proto messages.

### Does this PR introduce _any_ user-facing change?

A previously reverted perfomance improvement is now
put back in place.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #47667 from nija-at/log-debug.

Authored-by: Niranjan Jayakar <nija@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>