[SPARK-49137][SQL] When the Boolean condition in the `if statement` is invalid, an exception should be thrown
### What changes were proposed in this pull request?
The pr aims to throw an exception to the end-user when the `Boolean condition` in the `if statement` is unexpected, instead of quietly returning `false`.

### Why are the changes needed?
Reduce unexpected behavior when end-users make errors in writing Boolean statements.

### Does this PR introduce _any_ user-facing change?
Yes, when the Boolean condition in the if statement is illegal, an exception should be thrown instead of returning false directly.

### How was this patch tested?
Add new UT.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #47648 from panbingkun/SPARK-49137.

Authored-by: panbingkun <panbingkun@baidu.com>
Signed-off-by: Dongjoon Hyun <dhyun@apple.com>