[SPARK-49022] Use Column Node API in Column
### What changes were proposed in this pull request?
This PR makes the org.apache.spark.sql.Column and friends use the recently introduced ColumnNode API. This is a stepping stone towards making the Column API implementation agnostic.

Most of the changes are fairly mechanical, and they are mostly caused by the removal of the Column(Expression) constructor.

### Why are the changes needed?
We want to create unified Scala interface for Classic and Connect. A language agnostic Column API implementation is part of this.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Existing tests.

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #47688 from hvanhovell/SPARK-49022.

Authored-by: Herman van Hovell <herman@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>