[SPARK-49263][CONNECT] Spark Connect python client: Consistently handle boolean Dataframe reader options
### What changes were proposed in this pull request?

Using `spark.read.option("Foo", True)` resulted in an uppercase `'True'` string in Python Spark Connect client, while in all other cases (scala with both Spark Connect and no Spark Connect, pyspark with no Spark Connect) it would be normalized to `'true'`. This is because `to_str` helper should be used instead of `str`.

### Why are the changes needed?

This is now inconsistent with other cases. Passing `"True"` as boolean options seems to be breaking Delta CDF reader (to be fixed separately, that it should be able to handle the literal case-insensitively)

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Unittest added

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #47790 from juliuszsompolski/SPARK-49263.

Lead-authored-by: Hyukjin Kwon <gurwls223@apache.org>
Co-authored-by: Hyukjin Kwon <gurwls223@gmail.com>
Co-authored-by: Julek Sompolski <Juliusz Sompolski>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>