[SPARK-48796][SS] Load Column Family Id from RocksDBCheckpointMetadata for VCF when restarting
### What changes were proposed in this pull request?

Persisting the mapping between columnFamilyName to columnFamilyId in RocksDBCheckpointMetadata and RocksDBSnapshot. RocksDB will maintain an internal metadata of this mapping, and set this info on load. RocksDBStateStoreProvider can call columnFamily operations as usual, and RocksDB.scala will translate the name to the virtual column family ID.
### Why are the changes needed?

To enable the use of virtual column families, and the performance benefits it comes along with, with the TransformWithState operator

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Amended unit tests

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #47778 from ericm-db/vcf-integration-state-store.

Authored-by: Eric Marnadi <eric.marnadi@databricks.com>
Signed-off-by: Jungtaek Lim <kabhwan.opensource@gmail.com>