[SPARK-49269][SQL] Eagerly evaluate VALUES() list in AstBuilder
### What changes were proposed in this pull request?
This is a continuation of a prior performance improvement: https://github.com/apache/spark/pull/47428 that eagerly evaluates memory-heavy `UnresolvedUnlineTables` parse tree nodes as soon as they are constructed in the AstBuilder.

This PR applies this optimization to any statement that might contain one or more `VALUES()` clauses (such as subqueries etc), instead of just applying that optimization to `INSERT INTO ... VALUES` statements, which is what the prior PR did.

### Why are the changes needed?
With these changes we not only reduce the memory footprint of every statement that can contain the `VALUES()` clause, but we also improve upon the previous optimization as we avoid unnecessary traversals of the parse tree, which not only improves the runtime performance, but also minimizes the amount of time in which the `UnresolvedInlineTable` is kept in memory.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Provided scala tests.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #47791 from costas-db/eagerlyEvaluateUnresolvedInlineTableInAstBuilder.

Authored-by: Costas Zarifis <costas.zarifis@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>